Manage cache for BiocPkgTools

Managing user data is important to allow use of email functions such as `biocBuildEmail` and made easy with `BiocFileCache`.

setCache(
  directory = rappdirs::user_cache_dir("BiocPkgTools"),
  verbose = TRUE,
  ask = interactive()
)

pkgToolsCache(...)

Arguments

directory
The file location where the cache is located. Once set future downloads will go to this folder.

verbose
Whether to print descriptive messages

ask
logical (default TRUE when interactive session) Confirm the file location of the cache directory

...
For pkgToolsCache, arguments are passed to setCache

pkgToolsCache

Get the directory location of the cache. It will prompt the user to create a cache if not already created. A specific directory can be used via setCache.

setCache

Specify the directory location of the data cache. By default, it will got to the user's home/.cache and "appname" directory as specified by user_cache_dir. (default appname: pkgToolsCache)