Check the CRAN build report page and email a notification

The CRANstatus function allows users to check the status of a package and send an email report of any failures.

Usage

CRANstatus(
  pkg,
  core.name = NULL,
  core.email = NULL,
  core.id = NULL,
  to.mail = "maintainer@bioconductor.org",
  dry.run = TRUE,
  emailTemplate = templatePath("cranreport")
)

Arguments

pkg

character(1) The name of the package in trouble

core.name

character(1) The full name of the core team member

core.email

character(1) The Roswell Park email of the core team member

core.id

character(1) The internal identifier for the Roswell employee. This ID usually matches ^[A-Z]{2}[0-9]{5} for more recent identifiers.

to.mail

The email of the CRAN report recipient

dry.run

logical(1) Display the email without sending to the recipient. It only works for HTML email reports and ignored when textOnly=TRUE

emailTemplate

character(1) The path to the email template Rmd file as obtained by templatePath(). A custom template can be provided as file path.