Tidy Bioconductor build report results

The online Bioconductor build reports are great for humans to look at, but they are not easily computable. This function scrapes HTML and text files available from the build report online pages to generate a tidy data frame version of the build report.

Usage

biocBuildReport(
  version = BiocManager::version(),
  pkgType = c("software", "data-experiment", "data-annotation", "workflows"),
  stage.timings = FALSE
)

Arguments

version

character(1) the character version number as used to access the online build report. For example, "3.14". The default is the "current version" as given by BiocManager::version(). Note that this is a character vector of length one and not a number.

pkgType

character(1) The type of packages for which to get build status information for. Valid values are:

software: Software packages

data-experiment: Experiment data packages

data-annotation: Annotation data packages

workflows: Workflow packages

stage.timings

logical(1) Whether to include the start, end, and elapsed time for each build, check, install stage from each building in the result (default: FALSE)

Value

A tbl_df object with columns pkg, version, author, commit, date, node, stage, and result.

Examples

# Set the stage--what version of Bioc am I using?
BiocManager::version()
#> [1] ‘3.21’

latest_build <- biocBuildReport()
#> adding rname 'https://bioconductor.org/checkResults/3.21/bioc-LATEST/BUILD_STATUS_DB.txt'
#> adding rname 'https://bioconductor.org/checkResults/3.21/data-experiment-LATEST/BUILD_STATUS_DB.txt'
#> adding rname 'https://bioconductor.org/checkResults/3.21/data-annotation-LATEST/BUILD_STATUS_DB.txt'
#> adding rname 'https://bioconductor.org/checkResults/3.21/workflows-LATEST/BUILD_STATUS_DB.txt'
#> adding rname 'https://bioconductor.org/checkResults/3.21/bioc-LATEST/report.tgz'
#> adding rname 'https://bioconductor.org/checkResults/3.21/data-experiment-LATEST/report.tgz'
#> adding rname 'https://bioconductor.org/checkResults/3.21/data-annotation-LATEST/report.tgz'
#> adding rname 'https://bioconductor.org/checkResults/3.21/workflows-LATEST/report.tgz'
head(latest_build)
#> # A tibble: 6 × 12
#>   pkg    author  version git_last_commit git_last_commit_date pkgType Deprecated
#>   <chr>  <chr>   <chr>   <chr>           <dttm>               <chr>   <lgl>     
#> 1 ABSSeq Wentao… 1.62.0  df94993         2025-04-15 10:30:09  bioc    FALSE     
#> 2 ABSSeq Wentao… 1.62.0  df94993         2025-04-15 10:30:09  bioc    FALSE     
#> 3 ABSSeq Wentao… 1.62.0  df94993         2025-04-15 10:30:09  bioc    FALSE     
#> 4 ABSSeq Wentao… 1.62.0  df94993         2025-04-15 10:30:09  bioc    FALSE     
#> 5 ABSSeq Wentao… 1.62.0  df94993         2025-04-15 10:30:09  bioc    FALSE     
#> 6 ABSSeq Wentao… 1.62.0  df94993         2025-04-15 10:30:09  bioc    FALSE     
#> # ℹ 5 more variables: PackageStatus <chr>, node <chr>, stage <chr>,
#> #   result <chr>, bioc_version <chr>