Download and parse the build status information for Bioconductor packages

This function downloads and parses the build status information for Bioconductor packages. The build status information is available for the current release and the previous release. Other versions may be available.

biocBuildStatusDB(
  version = BiocManager::version(),
  pkgType = c("software", "data-experiment", "data-annotation", "workflows")
)

Arguments

version

character(1) The numeric version of Bioconductor to use, e.g., "3.19". Keywords "release" and "devel" are also accepted.

pkgType

character(1) The type of packages for which to get build status information for. Valid values are:

software: Software packages

data-experiment: Experiment data packages

data-annotation: Annotation data packages

workflows: Workflow packages

Value

A data.frame with the following columns:

pkg: The name of the package

node: The builder on which the package was built

stage: The stage of the build, e.g., 'install', 'buildsrc', 'checksrc', etc.

result: The status of the build, e.g., 'OK', 'ERROR', 'WARNINGS', etc.