get bioconductor download stats

get bioconductor download stats

biocDownloadStats()

Value

a data.frame of download stats for all bioconductor packages, in tidy format

Details

Note that bioconductor package download stats are not version-specific.

Examples

biocDownloadStats()

#> # A tibble: 418,288 x 6
#>    Package  Year Month Nb_of_distinct_IPs Nb_of_downloads repo    
#>    <fct>   <int> <fct>              <int>           <int> <chr>   
#>  1 ABarray  2019 Jan                  104             210 Software
#>  2 ABarray  2019 Feb                   80             164 Software
#>  3 ABarray  2019 Mar                   96             122 Software
#>  4 ABarray  2019 Apr                    0               0 Software
#>  5 ABarray  2019 May                    0               0 Software
#>  6 ABarray  2019 Jun                    0               0 Software
#>  7 ABarray  2019 Jul                    0               0 Software
#>  8 ABarray  2019 Aug                    0               0 Software
#>  9 ABarray  2019 Sep                    0               0 Software
#> 10 ABarray  2019 Oct                    0               0 Software
#> # … with 418,278 more rows