Bioconductor Maintained Packages

List all the packages associated with a maintainer. By default, it will return all packages associated with the maintainer@bioconductor.org email. hasBiocMaint returns a logical vector corresponding to the input character vector of packages indicating whether any package is maintained by the Bioconductor core team.

Usage

biocMaintained(
  main = "maintainer@bioconductor\\.org",
  version = BiocManager::version(),
  pkgType = c("software", "data-experiment", "workflows", "data-annotation")
)

hasBiocMaint(
  pkg,
  version = BiocManager::version(),
  main = "maintainer@bioconductor\\.org",
  repo = c("BioCsoft", "BioCexp", "BioCworkflows", "BioCann")
)

Arguments

main

character(1) The regex for searching through the Maintainer column as obtained from biocPkgList().

version

character(1) the character version number as used to access the online build report. For example, "3.14". The default is the "current version" as given by BiocManager::version(). Note that this is a character vector of length one and not a number.

pkgType

character(1) The type of packages for which to get build status information for. Valid values are:

software: Software packages

data-experiment: Experiment data packages

data-annotation: Annotation data packages

workflows: Workflow packages

pkg

character(1) A vector of package names (case sensitive).

repo

character() A vector of Bioconductor repositories to search through. By default, it will search through all Bioconductor repositories.

Value

For biocMaintained: a tibble of packages associated with the maintainer.

For hasBiocMaint: a logical vector indicating whether the package is maintained by Bioconductor.

Examples

biocMaintained()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
#> # A tibble: 37 × 47
#>    Package        Version Depends Suggests License MD5sum NeedsCompilation Title
#>    <chr>          <chr>   <list>  <list>   <chr>   <chr>  <chr>            <chr>
#>  1 annotate       1.86.0  <chr>   <chr>    Artist… 06361… no               "Ann…
#>  2 AnnotationDbi  1.70.0  <chr>   <chr>    Artist… 40203… no               "Man…
#>  3 AnnotationFil… 1.32.0  <chr>   <chr>    Artist… a4554… no               "Fac…
#>  4 AnnotationFor… 1.50.0  <chr>   <chr>    Artist… 00465… no               "Too…
#>  5 AnnotationHub  3.16.0  <chr>   <chr>    Artist… 8c2a0… yes              "Cli…
#>  6 AnnotationHub… 1.38.0  <chr>   <chr>    Artist… efdf1… no               "Tra…
#>  7 Biobase        2.68.0  <chr>   <chr>    Artist… c6039… yes              "Bio…
#>  8 BiocStyle      2.36.0  <chr>   <chr>    Artist… 53eaf… no               "Sta…
#>  9 BiocVersion    3.21.1  <chr>   <chr>    Artist… 58ff0… no               "Set…
#> 10 biocViews      1.76.0  <chr>   <chr>    Artist… f6464… no               "Cat…
#> # ℹ 27 more rows
#> # ℹ 39 more variables: Description <chr>, biocViews <list>, Author <list>,
#> #   Maintainer <list>, git_url <chr>, git_branch <chr>, git_last_commit <chr>,
#> #   git_last_commit_date <chr>, `Date/Publication` <chr>, source.ver <chr>,
#> #   win.binary.ver <chr>, `mac.binary.big-sur-x86_64.ver` <chr>,
#> #   `mac.binary.big-sur-arm64.ver` <chr>, vignettes <list>,
#> #   vignetteTitles <list>, hasREADME <chr>, hasNEWS <chr>, hasINSTALL <chr>, …

## maintained by Hervé and not maintainer at bioconductor dot org
hasBiocMaint("BiocGenerics")
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
#> BiocGenerics 
#>        FALSE