Get full Bioconductor software package listing, with details

The BiocViews-generated VIEWS file is available for Bioconductor release and devel repositories. It contains quite a bit more information from the package DESCRIPTION files than the PACKAGES file. In particular, it contains biocViews annotations and URLs for vignettes and developer URLs.

Usage

biocPkgList(
  version = BiocManager::version(),
  repo = c("BioCsoft", "BioCexp", "BioCworkflows", "BioCann", "CRAN"),
  addBiocViewParents = TRUE
)

Arguments

version

The requested Bioconductor version. Will default to use the BiocManager defaults (i.e., version()).

repo

character(1) The requested Bioconductor repository. The default is to pull from the "BioCsoft" repository. Possible repositories include "BioCsoft", "BioCexp", "BioCworkflows", "BioCann", and "CRAN". Note that not all repos are available for all versions, particularly older versions.

addBiocViewParents

logical(1) whether to add all biocViews parents to biocViews annotations.

Value

An object of class tbl_df.

Details

Since packages are annotated with the most specific views, the default functionality here is to add parent terms for all views for each package. For example, in the bioCsoft repository, all packages will have at least "Software" added to their biocViews. If one wants to stick to only the most specific terms, set addBiocViewParents to FALSE.

Examples

bpkgl <- biocPkgList(repo = "BioCsoft")
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
bpkgl
#> # A tibble: 2,341 × 47
#>    Package     Version Depends   Suggests  License MD5sum NeedsCompilation Title
#>    <chr>       <chr>   <list>    <list>    <chr>   <chr>  <chr>            <chr>
#>  1 a4          1.56.0  <chr [5]> <chr [7]> GPL-3   25af4… no               "Aut…
#>  2 a4Base      1.56.0  <chr [2]> <chr [4]> GPL-3   c8c01… no               "Aut…
#>  3 a4Classif   1.56.0  <chr [2]> <chr [4]> GPL-3   1f20f… no               "Aut…
#>  4 a4Core      1.56.0  <chr [1]> <chr [2]> GPL-3   533ce… no               "Aut…
#>  5 a4Preproc   1.56.0  <chr [1]> <chr [4]> GPL-3   852bc… no               "Aut…
#>  6 a4Reporting 1.56.0  <chr [1]> <chr [2]> GPL-3   83485… no               "Aut…
#>  7 ABarray     1.76.0  <chr [1]> <chr [2]> GPL     d5ce6… no               "Mic…
#>  8 abseqR      1.26.0  <chr [1]> <chr [1]> GPL-3 … d8996… no               "Rep…
#>  9 ABSSeq      1.62.0  <chr [2]> <chr [1]> GPL (>… c6765… no               "ABS…
#> 10 acde        1.38.0  <chr [2]> <chr [2]> GPL-3   4d7e4… no               "Art…
#> # ℹ 2,331 more rows
#> # ℹ 39 more variables: Description <chr>, biocViews <list>, Author <list>,
#> #   Maintainer <list>, git_url <chr>, git_branch <chr>, git_last_commit <chr>,
#> #   git_last_commit_date <chr>, `Date/Publication` <chr>, source.ver <chr>,
#> #   win.binary.ver <chr>, `mac.binary.big-sur-x86_64.ver` <chr>,
#> #   `mac.binary.big-sur-arm64.ver` <chr>, vignettes <list>,
#> #   vignetteTitles <list>, hasREADME <chr>, hasNEWS <chr>, hasINSTALL <chr>, …
unlist(bpkgl[1,'Depends'], use.names = FALSE)
#> [1] "a4Base"      "a4Preproc"   "a4Classif"   "a4Core"      "a4Reporting"

# Get a list of all packages that
# import "GEOquery"
library(dplyr)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
bpkgl |>
  filter(Package == 'GEOquery') |>
  pull('importsMe') |>
  unlist()
#>  [1] "bigmelon"                       "ChIPXpress"                    
#>  [3] "DExMA"                          "EGAD"                          
#>  [5] "GEOexplorer"                    "minfi"                         
#>  [7] "Moonlight2R"                    "MoonlightR"                    
#>  [9] "phantasus"                      "recount"                       
#> [11] "BeadArrayUseCases"              "BioPlex"                       
#> [13] "GSE13015"                       "healthyControlsPresenceChecker"
#> [15] "geneExpressionFromGEO"          "RCPA"