Tidy Bioconductor build report results

The online Bioconoductor build reports are great for humans to look at, but they are not easily computable. This function simply scrapes HTML and text files available from the build report online pages to generate a tidy data frame version of the build report.

bioc_build_report(version = as.character(biocVersion()))

Arguments

version
character(1) the version number as used to access the online build report. For example, "3.6".

Value

a data.frame with columns pkg, version, author, commit, date, node, stage, and result.

Examples

latest_build = bioc_build_report()

#> Joining, by = "pkg"

head(latest_build)

#>   pkg version         author   commit last_changed_date    node    stage
#> 1  a4  1.27.0 Tobias Verbeke  52d6c2b        2017-10-30 malbec2  install
#> 2  a4  1.27.0 Tobias Verbeke  52d6c2b        2017-10-30 malbec2 buildsrc
#> 3  a4  1.27.0 Tobias Verbeke  52d6c2b        2017-10-30 malbec2 checksrc
#> 4  a4  1.27.0 Tobias Verbeke  52d6c2b        2017-10-30  tokay2  install
#> 5  a4  1.27.0 Tobias Verbeke  52d6c2b        2017-10-30  tokay2 buildsrc
#> 6  a4  1.27.0 Tobias Verbeke  52d6c2b        2017-10-30  tokay2 checksrc
#>      result bioc_version
#> 1 NotNeeded          3.7
#> 2        OK          3.7
#> 3        OK          3.7
#> 4 NotNeeded          3.7
#> 5     ERROR          3.7
#> 6   skipped          3.7