Work with Bioconductor package dependencies

Bioconductor is built using an extensive set of core capabilities and data structures. This leads to package developers depending on other packages for interoperability and functionality. This function extracts package dependency information from biocPkgList and returns a tidy data.frame that can be used for analysis and to build graph structures of package dependencies.

buildPkgDependencyDataFrame(
  dependencies = c("Depends", "Imports", "Suggests"),
  ...
)

Arguments

dependencies

character() vector including one or more of "Depends", "Imports", or "Suggests". Default is to include all possibilities.

...

parameters passed along to biocPkgList

Value

A data.frame (also a tbl_df) of S3 class "biocDepDF" including columns "Package", "dependency", and "edgetype".

Note

This function requires network access.

See also

Examples

# performs a network call, so must be online.
library(BiocPkgTools)
depdf = buildPkgDependencyDataFrame()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
head(depdf)
#> # A tibble: 6 × 3
#>   Package dependency  edgetype
#>   <chr>   <chr>       <chr>   
#> 1 a4      a4Base      Depends 
#> 2 a4      a4Preproc   Depends 
#> 3 a4      a4Classif   Depends 
#> 4 a4      a4Core      Depends 
#> 5 a4      a4Reporting Depends 
#> 6 a4Base  a4Preproc   Depends 
library(dplyr)
# filter to include only "Imports" type
# dependencies
imports_only = depdf %>% filter(edgetype=='Imports')

# top ten most imported packages
imports_only %>% select(dependency) %>%
  group_by(dependency) %>% tally() %>%
  arrange(desc(n))
#> # A tibble: 1,607 × 2
#>    dependency       n
#>    <chr>        <int>
#>  1 stats         1092
#>  2 methods        986
#>  3 utils          899
#>  4 graphics       578
#>  5 ggplot2        548
#>  6 grDevices      536
#>  7 S4Vectors      504
#>  8 IRanges        396
#>  9 BiocGenerics   347
#> 10 dplyr          344
#> # … with 1,597 more rows

# Bioconductor packages doing the largest
# amount of importing
largest_importers = imports_only %>%
  select(Package) %>%
  group_by(Package) %>% tally() %>%
  arrange(desc(n))

# not sure what these packages do. Join
# to their descriptions
biocPkgList() %>% select(Package, Description) %>%
  left_join(largest_importers) %>% arrange(desc(n)) %>%
  head()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> Joining, by = "Package"
#> # A tibble: 6 × 3
#>   Package      Description                                                     n
#>   <chr>        <chr>                                                       <int>
#> 1 singleCellTK "Run common single cell analysis in the R console or\ndire…    73
#> 2 DAPAR        "This package contains a collection of functions for the\n…    56
#> 3 ChromSCape   "ChromSCape - Chromatin landscape profiling for Single\nCe…    54
#> 4 musicatk     "Mutational signatures are carcinogenic exposures or\naber…    54
#> 5 metaseqR2    "Provides an interface to several normalization and\nstati…    51
#> 6 ELMER        "ELMER is designed to use DNA methylation and gene\nexpres…    45