When did a package enter Bioconductor?

This function uses the biocDownloadStats data to approximate when a package entered Bioconductor. Note that the download stats go back only to 2009.

Usage

firstInBioc(download_stats)

Arguments

download_stats

a data.frame from biocDownloadStats

Examples

dls <- biocDownloadStats()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
tail(firstInBioc(dls))
#> # A tibble: 6 × 7
#> # Groups:   Package [6]
#>   pkgType      Package  Year Month Nb_of_distinct_IPs Nb_of_downloads Date      
#>   <chr>        <chr>   <int> <chr>              <int>           <int> <date>    
#> 1 data-annota… ygs98f…  2013 Jan                    0               0 2013-01-01
#> 2 data-annota… ygs98p…  2009 Jan                   25              43 2009-01-01
#> 3 data-annota… zebraf…  2009 Jan                   37              57 2009-01-01
#> 4 data-annota… zebraf…  2009 Jan                   23              40 2009-01-01
#> 5 data-annota… zebraf…  2009 Jan                   67              94 2009-01-01
#> 6 data-annota… zebraf…  2009 Jan                   31              51 2009-01-01