Calculate the years in Bioconductor

This function determines the number of years a package has been in Bioconductor. Available information includes first Bioconductor version a package appeared and the current length of time in Bioconductor. If a package has been removed from Bioconductor, information on the last Bioconductor version and approximate time in Bioconductor before removal is available.

Usage

getPkgYearsInBioc(pkglist = NULL)

Arguments

pkglist

List of packages to retrieve information. If default NULL, returns a tibble of all Bioconductor packages.

Value

'tibble' with the following columns:

package: name of Bioconductor package

category: bioc, data/experiment, data/annotation, workflow

first_version_available: Bioconductor version (e.g. 1.9, 3.21) the package first became available

first_version_release_date: Equivalent calendar date of given Bioconductor release

approx_years_in: Numeric indicator of years in Bioconductor. If empty, indicates package was removed. See final three columns for more information.

last_version_available: If package was removed from Bioconductor, the last Bioconductor version (e.g. 1.9, 3.21) the package was able to be installed

last_version_release_date: Equivalent calendar date of given Bioconductor release

years_before_rm: If removed, how many years it was in Bioconductor

Author

Lori Shepherd Kern, Robert Shear

Examples

if (FALSE) { # \dontrun{
   ## full table all Bioconductor packages
   tbl <- getPkgYearsInBioc()

   ## example of package list. Packages active in Bioconductor
   tbl <- getPkgYearsInBioc(c("BiocFileCache", "BiocPkgTools"))

   ## example of a package that has been removed from Bioconductor
   tbl <- getPkgYearsInBioc("ensemblVEP")
} # }