Obtain all the reverse dependencies for a Bioconductor package

The function returns a slightly upgraded list with dependency types as elements and package names in each of those elements, if any. The types of dependencies can be seen in the which argument documentation.

Usage

pkgBiocRevDeps(
  pkg,
  pkgType = c("software", "data-experiment", "workflows", "data-annotation"),
  which = "all",
  only.bioc = TRUE,
  version = BiocManager::version()
)

# S3 method for class 'biocrevdeps'
summary(object, ...)

Arguments

pkg

character(1) The package for which to look up dependencies.

pkgType

character() Any of 'software', 'data-experiment', 'workflows', and / or 'data-annotation' (defaults to all)

which

a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

only.bioc

logical(1) Whether to only return Bioconductor dependencies in the list (default TRUE)

version

(Optional) character(1) or package_version indicating the Bioconductor version (e.g., "3.8") for which repositories are required.

object

an object for which a summary is desired.

...

additional arguments affecting the summary produced.

Value

A biocrevdeps list class object

Details

The summary method of the biocrevdeps class given by pkgBiocRevDeps provides a tally in each dependency field.

Examples

rdeps <- pkgBiocRevDeps("MultiAssayExperiment", which = "all")
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
rdeps
#> $Depends
#>  [1] "alabaster.mae"        "CAGEr"                "cBioPortalData"      
#>  [4] "ClassifyR"            "evaluomeR"            "hipathia"            
#>  [7] "HoloFoodR"            "InTAD"                "MGnifyR"             
#> [10] "mia"                  "midasHLA"             "MIRit"               
#> [13] "missRows"             "QFeatures"            "RFLOMICS"            
#> [16] "terraTCGAdata"        "curatedPCaData"       "curatedTCGAData"     
#> [19] "microbiomeDataSets"   "OMICsPCAdata"         "scMultiome"          
#> [22] "SingleCellMultiModal"
#> 
#> $Imports
#>  [1] "AffiXcan"               "AMARETTO"               "animalcules"           
#>  [4] "autonomics"             "biosigner"              "CoreGx"                
#>  [7] "corral"                 "ELMER"                  "FindIT2"               
#> [10] "gDRcore"                "gDRimport"              "gDRutils"              
#> [13] "gINTomics"              "glmSparseNet"           "GOpro"                 
#> [16] "hermes"                 "Lheuristic"             "LinkHD"                
#> [19] "metabolomicsWorkbenchR" "MOMA"                   "MOSClip"               
#> [22] "msqrob2"                "MuData"                 "MultiBaC"              
#> [25] "MultimodalExperiment"   "nipalsMCIA"             "OMICsPCA"              
#> [28] "omicsPrint"             "omXplore"               "padma"                 
#> [31] "PDATK"                  "PharmacoGx"             "phenomis"              
#> [34] "ropls"                  "scp"                    "scPipe"                
#> [37] "survClust"              "TCGAutils"              "TENET"                 
#> [40] "vsclust"                "xcore"                  "curatedTBData"         
#> [43] "HMP2Data"               "LegATo"                 "MetaScope"             
#> [46] "TENET.ExperimentHub"   
#> 
#> $LinkingTo
#> character(0)
#> 
#> $Suggests
#>  [1] "BiocGenerics"     "CNVRanger"        "funOmics"         "maftools"        
#>  [5] "MOFA2"            "MultiDataSet"     "RaggedExperiment" "updateObject"    
#>  [9] "brgedata"         "MOFAdata"        
#> 
#> $Enhances
#> character(0)
#> 
#> attr(,"package")
#> [1] "MultiAssayExperiment"
#> attr(,"class")
#> [1] "biocrevdeps"
summary(rdeps)
#>                      Depends Imports LinkingTo Suggests Enhances Total
#> MultiAssayExperiment      22      46         0       10        0    78