Calculate dependency gain achieved by excluding combinations of packages

Calculate dependency gain achieved by excluding combinations of packages

Usage

pkgCombDependencyGain(pkg, depdf, maxNbr = 3L)

Arguments

pkg

character, the name of the package for which we want to estimate the dependency gain

depdf

a tidy data frame with package dependency information obtained through the function buildPkgDependencyDataFrame

maxNbr

numeric, the maximal number of direct dependencies to leave out simultaneously

Value

A data frame with three columns: ExclPackages (the excluded direct dependencies), NbrExcl (the number of excluded direct dependencies), DepGain (the dependency gain from excluding these direct dependencies)

Author

Charlotte Soneson

Examples

depdf <- buildPkgDependencyDataFrame(
  dependencies=c("Depends", "Imports"), 
  repo=c("BioCsoft", "CRAN")
)
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
pcd <- pkgCombDependencyGain('GEOquery', depdf, maxNbr = 3L)
head(pcd[order(pcd$DepGain, decreasing = TRUE), ])
#>                                   Packages NbrExcl DepGain
#> 337   SummarizedExperiment, R.utils, readr       3      27
#> 247 S4Vectors, SummarizedExperiment, readr       3      26
#> 348     SummarizedExperiment, limma, readr       3      26
#> 375     SummarizedExperiment, httr2, readr       3      26
#> 383   SummarizedExperiment, readr, rentrez       3      26
#> 384     SummarizedExperiment, readr, rvest       3      26