Report package imported functionality

Function adapted from 'itdepends::dep_usage_pkg' at https://github.com/r-lib/itdepends to obtain the functionality imported and used by a given package.

pkgDepImports(pkg)

Arguments

pkg

character() name of the package for which we want to obtain the functionality calls imported from its dependencies and used within the package.

Value

A tidy data frame with two columns:

pkg: name of the package dependency.

fun: name of the functionality call imported from the the dependency in the column pkg and used within the analyzed package.

Details

Certain imported elements, such as built-in constants, will not be identified as imported functionality by this function.

Author

Robert Castelo

Examples

pkgDepImports('BiocPkgTools')
#> # A tibble: 52 × 2
#>    pkg         fun               
#>    <chr>       <chr>             
#>  1 BiocManager version           
#>  2 DT          datatable         
#>  3 RBGL        transitive.closure
#>  4 dplyr       bind_rows         
#>  5 dplyr       collect           
#>  6 dplyr       desc              
#>  7 dplyr       filter            
#>  8 dplyr       group_by          
#>  9 dplyr       mutate            
#> 10 dplyr       rename            
#> # … with 42 more rows