What is a package's download rank?

This function uses available.packages to calculate the download rank percentile of a given package. It approximates what is observed in the Bioconductor landing page.

pkgDownloadRank(
  pkg,
  pkgType = c("software", "data-experiment", "workflows", "data-annotation"),
  version = BiocManager::version()
)

Arguments

pkg

character(1) The name of a Bioconductor package

pkgType

character(1) Either one of 'software', 'data-experiment', 'workflows', or 'data-annotation' (defaults to 'all' or 'software')

version

(Optional) character(1) or package_version indicating the Bioconductor version (e.g., "3.8") for which repositories are required.

Value

The package's percentile rank, in terms of download statistics, and proportion in the name

Examples

## Percentile rank for BiocGenerics (top 1%)
pkgDownloadRank("BiocGenerics", "software")
#> 1/2123 
#>   0.05