Bioconductor Binary Repository Statistics

Summarize binary packages compatible with the Bioconductor or Terra container in use.

Usage

repositoryStats(
  version = BiocManager::version(),
  binary_repository = BiocManager::containerRepository(version)
)

# S3 method for class 'repositoryStats'
print(x, ...)

Arguments

version

(Optional) character(1) or package_version indicating the Bioconductor version (e.g., "3.8") for which repositories are required.

binary_repository

character(1) location of binary repository as given by BiocManager::containerRepository (default)

x

the object returned by repositoryStats().

...

further arguments passed to or from other methods (not used).

Value

a list of class repositoryStats with the following fields:

container: character(1) container label, e.g., bioconductor_docker, or NA if not evaluated on a supported container

bioconductor_version: package_version the Bioconductor version provided by the user.

repository_exists: logical(1) TRUE if a binary repository exists for the container and Bioconductor_Version version.

bioconductor_binary_repository: character(1) repository location, if available, or NA if the repository does not exist.

n_software_packages: integer(1) number of software packages in the Bioconductor source repository.

n_binary_packages: integer(1) number of binary packages available. When a binary repository exists, this number is likely to be larger than the number of source software packages, because it includes the binary version of the source software packages, as well as the (possibly CRAN) dependencies of the binary packages

n_binary_software_packages: integer(1) number of binary packages derived from Bioconductor source packages. This number is less than or equal to n_software_packages.

missing_binaries: integer(1) the number of Bioconductor source software packages that are not present in the binary repository.

out_of_date_binaries: integer(1) the number of Bioconductor source software packages that are newer than their binary counterpart. A newer source software package might occur when the main Bioconductor build system has updated a package after the most recent run of the binary build system.

Methods (by generic)

print(repositoryStats): Print a summary of package availability in binary repositories.

Author

M. Morgan

Examples

stats <- repositoryStats() # obtain statistics
stats                       # display a summary
#> Container: NA
#> Bioconductor version: 3.21
#> Bioconductor binary repos:  NA
#> PACKAGES timestamp: NA
#> Query timestamp: 2025-05-12 15:01 UTC
#> Bioconductor software packages: 2308
#> Binary packages: 0
#> Binary software packages: 0
#> Missing binary software packages: 2308
#> Out-of-date binary software packages: 0
stats$container             # access an element for further computation
#> [1] NA