Subset graph by degree

While the inducedSubgraphByPkgs returns the subgraph with the minimal connections between named packages, this function takes a vector of package names, a degree (1 or more) and returns the subgraph(s) that are within degree of the package named.

subgraphByDegree(g, pkg, degree = 1, ...)

Arguments

g

an igraph graph, typically created by buildPkgDependencyIgraph

pkg

character(1) package name from which to measure degree.

degree

integer(1) degree, limit search for adjacent vertices to this degree.

...

passed on to distances

Value

An igraph graph, with only nodes and their edges within degree of the named package

Examples

g = buildPkgDependencyIgraph(buildPkgDependencyDataFrame())
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
g2 = subgraphByDegree(g, 'GEOquery')
g2
#> IGRAPH e2c4e23 DN-- 63 288 -- 
#> + attr: name (v/c), edgetype (e/c)
#> + edges from e2c4e23 (vertex names):
#>  [1] autonomics  ->BiocGenerics autonomics  ->GEOquery    
#>  [3] autonomics  ->limma        autonomics  ->methods     
#>  [5] autonomics  ->magrittr     autonomics  ->data.table  
#>  [7] autonomics  ->rmarkdown    autonomics  ->knitr       
#>  [9] autonomics  ->tidyr        autonomics  ->R.utils     
#> [11] autonomics  ->testthat     bigmelon    ->Biobase     
#> [13] bigmelon    ->BiocGenerics bigmelon    ->BiocGenerics
#> [15] bigmelon    ->GEOquery     bigmelon    ->minfi       
#> + ... omitted several edges