What are the issues, pulls, commits created since a date?

This function uses the gh package to get a list of either issues, pull requests, or GitHub commits since the specified date for a particular GitHub repository. The repository must have both the username / organization and the name, e.g., "Bioconductor/S4Vectors".

activitySince(
  gh_repo,
  activity = c("issues", "pulls", "commits"),
  status = c("closed", "open", "all"),
  Date,
  issue_metadata = c("created_at", "number", "title"),
  token = NULL
)

Arguments

gh_repo

character(1) The GitHub repository location including the username / organization and the repository name, e.g., "Bioconductor/S4Vectors"

activity

character(1) The type of repository activity to pull from the GitHub API. It can be one of "issues" (default), "pulls", or "commits".

status

character(1) One of 'closed', 'open', or 'all' corresponding to the issue state desired from the GitHub API (Default: "closed"). This argument is ignored for the "commits" activity report.

Date

character(1) The date cutoff from which to analyze closed issues in the YYYY-MM-DD or YYYY-MM-DDTHH:MM:SSZ format (ISO 8601).

issue_metadata

character() The metadata labels to extract from the gh::gh response. See ?gh::gh for more details. Defaults to 'created_at', 'number', and 'title'. This argument is ignored for the "commits" activity report.

token

character(1) For big requests, e.g., commit history, you may be prompted to use a GitHub Personal Access Token. Enter the token as plain text.

Value

A tibble with three columns corresponding to issue metadata (i.e., "created_at", "number", "title")

Details

The tibble returned by the commits activity report contains five columns:

'committer_date'

'commit' - hash

'parents' - hash of parent for merge commits

'author'

'message'

For information on other columns, refer to the GitHub API under repository issues or pulls (e.g., /repos/:repo/issues).

Examples

activitySince("Bioconductor/S4Vectors", "issues", "closed", "2021-05-01")
#> # A tibble: 28 × 3
#>    created_at           number title                                            
#>    <chr>                 <int> <chr>                                            
#>  1 2022-09-09T15:27:30Z    109 "show,DataFrame throws error on DataFrame with n…
#>  2 2022-06-07T21:33:36Z    108 "DataFrame assignment forces names of other colu…
#>  3 2022-05-26T09:53:21Z    107 "Improve and expand the Factor-class with raw in…
#>  4 2022-05-03T04:52:33Z    106 "Could I get old version S4Vectors?"             
#>  5 2022-04-05T17:41:19Z    105 "  'class2' must be the name of a class or a cla…
#>  6 2022-04-01T04:38:23Z    104 "Output class of the median of an Rle changed"   
#>  7 2022-03-31T22:56:31Z    103 "`List` constructor fails for freshly unserializ…
#>  8 2022-03-27T21:42:51Z    102 "combineCols() error"                            
#>  9 2022-03-22T17:08:47Z    101 "Error: duplicate row.names"                     
#> 10 2022-03-04T17:23:51Z    100 "Re-assignment of a nested `Dataframe` via `[,<-…
#> # … with 18 more rows

activitySince("Bioconductor/S4Vectors", "issues", "open", "2022-05-01")
#> # A tibble: 0 × 0

activitySince("Bioconductor/S4Vectors", "commits", Date = "2022-05-01")
#> # A tibble: 6 × 5
#>   committer_date       commit                             parents author message
#>   <chr>                <chr>                              <chr>   <chr>  <chr>  
#> 1 2022-09-16T18:57:10Z 27c5dc4f20fdb9b14797b50a578b85e7d… NA      Hervé… "Versi…
#> 2 2022-09-16T18:48:19Z 78d58b39ba52a5992397e0d628b8611c5… NA      Aaron… "Expan…
#> 3 2022-08-31T19:51:14Z 27c2091ab3e8a62ec7020d99ce56f4924… NA      Hervé… "Add d…
#> 4 2022-08-29T19:09:14Z e3184a14cf57cabf856b3ac402d84fdd2… NA      Hervé… "small…
#> 5 2022-08-29T18:32:05Z bcd7c75a98f5776bd312bcafb3a50e50a… NA      Hervé… "Store…
#> 6 2022-06-08T02:36:53Z c8d9af3a53f0eadd0d5e609898d74fa69… NA      Hervé… "Subse…