Get download statistics for Bioconductor packages distributed via Anaconda.

Get download statistics for Bioconductor packages distributed via Anaconda.

anacondaDownloadStats()

Value

A data.frame of download statistics for all Bioconductor packages distributed by Anaconda, in tidy format. Note: Anaconda do not provide counts for unique IP addresses. This column is listed as NA for all packages to provide continuity with data from Bioconductor.org obtained by biocDownloadStats. The counts are updated monthly, so do not expect to see counts for the current month.

Details

Anaconda provide daily download counts for all software packages they distribute. These are summarised into monthly tables of counts and made available from https://github.com/grimbough/anaconda-download-stats This function provides a mechanism to download these monthly counts for Bioconductor packages distributed through Anaconda.

Author

Mike L. Smith

Examples

anacondaDownloadStats()
#> # A tibble: 285,900 × 7
#>    Package Year  Month Nb_of_distinct_IPs Nb_of_downloads repo     Date      
#>    <chr>   <chr> <chr>              <int>           <dbl> <chr>    <date>    
#>  1 ABAData 2018  Apr                   NA               8 Anaconda 2018-04-01
#>  2 ABAData 2018  Aug                   NA               5 Anaconda 2018-08-01
#>  3 ABAData 2018  Dec                   NA             133 Anaconda 2018-12-01
#>  4 ABAData 2018  Jul                   NA               6 Anaconda 2018-07-01
#>  5 ABAData 2018  Jun                   NA              18 Anaconda 2018-06-01
#>  6 ABAData 2018  Mar                   NA              13 Anaconda 2018-03-01
#>  7 ABAData 2018  May                   NA               6 Anaconda 2018-05-01
#>  8 ABAData 2018  Nov                   NA             138 Anaconda 2018-11-01
#>  9 ABAData 2018  Oct                   NA              33 Anaconda 2018-10-01
#> 10 ABAData 2018  Sep                   NA               5 Anaconda 2018-09-01
#> # ℹ 285,890 more rows