Tidy Bioconductor build report results

The online Bioconoductor build reports are great for humans to look at, but they are not easily computable. This function scrapes HTML and text files available from the build report online pages to generate a tidy data frame version of the build report.

biocBuildReport(version = BiocManager::version(), stage.timings = FALSE)

Arguments

version

character(1) the character version number as used to access the online build report. For example, "3.14". The default is the "current version" as given by BiocManager::version(). Note that this is a character vector of length one and not a number.

stage.timings

logical(1) Whether to include the start, end, and elapsed time for each build, check, install stage from each building in the result (default: FALSE)

Value

A tbl_df object with columns pkg, version, author, commit, date, node, stage, and result.

Examples

# Set the stage--what version of Bioc am I using?
BiocManager::version()
#> [1] ‘3.14’

latest_build <- biocBuildReport()
head(latest_build)
#> # A tibble: 6 × 11
#>   pkg           author    version git_last_commit git_last_commit_da… Deprecated
#>   <chr>         <chr>     <chr>   <chr>           <dttm>              <lgl>     
#> 1 ABAEnrichment Steffi G… 1.24.0  5d20752         2021-10-26 12:22:01 FALSE     
#> 2 ABAEnrichment Steffi G… 1.24.0  5d20752         2021-10-26 12:22:01 FALSE     
#> 3 ABAEnrichment Steffi G… 1.24.0  5d20752         2021-10-26 12:22:01 FALSE     
#> 4 ABAEnrichment Steffi G… 1.24.0  5d20752         2021-10-26 12:22:01 FALSE     
#> 5 ABAEnrichment Steffi G… 1.24.0  5d20752         2021-10-26 12:22:01 FALSE     
#> 6 ABAEnrichment Steffi G… 1.24.0  5d20752         2021-10-26 12:22:01 FALSE     
#> # … with 5 more variables: PackageStatus <chr>, node <chr>, stage <chr>,
#> #   result <chr>, bioc_version <chr>