get bioconductor download stats

get bioconductor download stats

biocDownloadStats()

Value

a data.frame of download stats for all bioconductor packages, in tidy format

Details

Note that bioconductor package download stats are not version-specific.

Examples

biocDownloadStats()

#> # A tibble: 464,633 x 7
#>    Package  Year Month Nb_of_distinct_IPs Nb_of_downloads repo     Date      
#>    <chr>   <int> <chr>              <int>           <int> <chr>    <date>    
#>  1 ABarray  2020 Jan                   56             117 Software 2020-01-01
#>  2 ABarray  2020 Feb                   52             126 Software 2020-02-01
#>  3 ABarray  2020 Mar                   23              42 Software 2020-03-01
#>  4 ABarray  2020 Apr                    0               0 Software 2020-04-01
#>  5 ABarray  2020 May                    0               0 Software 2020-05-01
#>  6 ABarray  2020 Jun                    0               0 Software 2020-06-01
#>  7 ABarray  2020 Jul                    0               0 Software 2020-07-01
#>  8 ABarray  2020 Aug                    0               0 Software 2020-08-01
#>  9 ABarray  2020 Sep                    0               0 Software 2020-09-01
#> 10 ABarray  2020 Oct                    0               0 Software 2020-10-01
#> # … with 464,623 more rows