Get Bioconductor download statistics

Get Bioconductor download statistics

biocDownloadStats(
  pkgType = c("all", "software", "data-experiment", "workflows", "data-annotation")
)

Arguments

pkgType

character(1) Either one of 'software', 'data-experiment', 'workflows', or 'data-annotation' (defaults to 'all' or 'software')

Value

A data.frame of download statistics for all Bioconductor packages, in tidy format

Details

Note that Bioconductor package download stats are not version-specific.

Examples

biocDownloadStats()
#> # A tibble: 526,080 × 7
#>    pkgType  Package  Year Month Nb_of_distinct_IPs Nb_of_downloads Date      
#>    <chr>    <chr>   <int> <chr>              <int>           <int> <date>    
#>  1 software ABarray  2022 Jan                   24              30 2022-01-01
#>  2 software ABarray  2022 Feb                    0               0 2022-02-01
#>  3 software ABarray  2022 Mar                    0               0 2022-03-01
#>  4 software ABarray  2022 Apr                    0               0 2022-04-01
#>  5 software ABarray  2022 May                    0               0 2022-05-01
#>  6 software ABarray  2022 Jun                    0               0 2022-06-01
#>  7 software ABarray  2022 Jul                    0               0 2022-07-01
#>  8 software ABarray  2022 Aug                    0               0 2022-08-01
#>  9 software ABarray  2022 Sep                    0               0 2022-09-01
#> 10 software ABarray  2022 Oct                    0               0 2022-10-01
#> # … with 526,070 more rows