get bioconductor download stats

get bioconductor download stats

biocDownloadStats()

Value

a data.frame of download stats for all bioconductor packages, in tidy format

Details

Note that bioconductor package download stats are not version-specific.

Examples

biocDownloadStats()

#> # A tibble: 376,597 x 6
#>    Package  Year Month Nb_of_distinct_IPs Nb_of_downloads repo    
#>    <fct>   <int> <fct>              <int>           <int> <chr>   
#>  1 ABarray  2018 Jan                  117             150 Software
#>  2 ABarray  2018 Feb                   97             125 Software
#>  3 ABarray  2018 Mar                  102             121 Software
#>  4 ABarray  2018 Apr                  229             359 Software
#>  5 ABarray  2018 May                   99             134 Software
#>  6 ABarray  2018 Jun                  133             209 Software
#>  7 ABarray  2018 Jul                  107             143 Software
#>  8 ABarray  2018 Aug                   68              98 Software
#>  9 ABarray  2018 Sep                   90             142 Software
#> 10 ABarray  2018 Oct                   13              24 Software
#> # ... with 376,587 more rows