Get Bioconductor download statistics

Get Bioconductor download statistics

biocDownloadStats(
  pkgType = c("software", "data-experiment", "workflows", "data-annotation")
)

Arguments

pkgType

character() All, some, or one of 'software', 'data-experiment', 'workflows', or 'data-annotation' (defaults to all types)

Value

A tibble of download statistics for all Bioconductor packages

Details

Note that Bioconductor package download stats are not version-specific.

Examples

biocDownloadStats()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
#> adding rname 'https://bioconductor.org/packages/stats/bioc/bioc_pkg_stats.tab'
#> adding rname 'https://bioconductor.org/packages/stats/data-experiment/experiment_pkg_stats.tab'
#> adding rname 'https://bioconductor.org/packages/stats/workflows/workflows_pkg_stats.tab'
#> adding rname 'https://bioconductor.org/packages/stats/data-annotation/annotation_pkg_stats.tab'
#> # A tibble: 457,272 × 7
#>    pkgType  Package  Year Month Nb_of_distinct_IPs Nb_of_downloads Date      
#>    <chr>    <chr>   <int> <chr>              <int>           <int> <date>    
#>  1 software a4       2025 Jan                  171             628 2025-01-01
#>  2 software a4       2025 Feb                  215             532 2025-02-01
#>  3 software a4       2025 Mar                  189             462 2025-03-01
#>  4 software a4       2025 Apr                  157             414 2025-04-01
#>  5 software a4       2025 May                    0               0 2025-05-01
#>  6 software a4       2025 Jun                    0               0 2025-06-01
#>  7 software a4       2025 Jul                    0               0 2025-07-01
#>  8 software a4       2025 Aug                    0               0 2025-08-01
#>  9 software a4       2025 Sep                    0               0 2025-09-01
#> 10 software a4       2025 Oct                    0               0 2025-10-01
#> # ℹ 457,262 more rows