Bioconductor Maintained Packages

List all the packages associated with a maintainer. By default, it will return all packages associated with the maintainer@bioconductor.org email.

biocMaintained(
  version = BiocManager::version(),
  main = "maintainer@bioconductor\\.org"
)

Arguments

version

character(1) the character version number as used to access the online build report. For example, "3.14". The default is the "current version" as given by BiocManager::version(). Note that this is a character vector of length one and not a number.

main

character(1) The regex for searching through the Maintainer column as obtained from biocPkgList().

Examples

biocMaintained()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> # A tibble: 53 × 46
#>    Package  Version Depends Sugge…¹ License MD5sum Needs…² Title Descr…³ biocV…⁴
#>    <chr>    <chr>   <list>  <list>  <chr>   <chr>  <chr>   <chr> <chr>   <list> 
#>  1 annotate 1.74.0  <chr>   <chr>   Artist… 9ea0f… no      "Ann… "Using… <chr>  
#>  2 Annotat… 1.58.0  <chr>   <chr>   Artist… 0ca49… no      "Man… "Imple… <chr>  
#>  3 Annotat… 1.20.0  <chr>   <chr>   Artist… 02a5a… no      "Fac… "This … <chr>  
#>  4 Annotat… 1.38.1  <chr>   <chr>   Artist… 903b1… no      "Too… "Provi… <chr>  
#>  5 Annotat… 3.4.0   <chr>   <chr>   Artist… 51942… yes     "Cli… "This … <chr>  
#>  6 Annotat… 1.26.1  <chr>   <chr>   Artist… 7af52… no      "Tra… "These… <chr>  
#>  7 Biobase  2.56.0  <chr>   <chr>   Artist… 244f7… yes     "Bio… "Funct… <chr>  
#>  8 BiocChe… 1.32.1  <chr>   <chr>   Artist… e5b71… no      "Bio… "BiocC… <chr>  
#>  9 BiocDoc… 1.8.1   <chr>   <chr>   Artist… 72832… no      "Acc… "Packa… <chr>  
#> 10 BiocGen… 0.42.0  <chr>   <chr>   Artist… 7e10e… no      "S4 … "The p… <chr>  
#> # … with 43 more rows, 36 more variables: Author <list>, Maintainer <list>,
#> #   git_url <chr>, git_branch <chr>, git_last_commit <chr>,
#> #   git_last_commit_date <chr>, `Date/Publication` <chr>, source.ver <chr>,
#> #   win.binary.ver <chr>, mac.binary.ver <chr>, vignettes <list>,
#> #   vignetteTitles <list>, hasREADME <chr>, hasNEWS <chr>, hasINSTALL <chr>,
#> #   hasLICENSE <chr>, Rfiles <list>, dependencyCount <chr>, Imports <list>,
#> #   Enhances <list>, dependsOnMe <list>, suggestsMe <list>, …