Bioconductor Maintained Packages

List all the packages associated with a maintainer. By default, it will return all packages associated with the maintainer@bioconductor.org email.

biocMaintained(
  version = BiocManager::version(),
  main = "maintainer@bioconductor\\.org"
)

Arguments

version

character(1) the character version number as used to access the online build report. For example, "3.14". The default is the "current version" as given by BiocManager::version(). Note that this is a character vector of length one and not a number.

main

character(1) The regex for searching through the Maintainer column as obtained from biocPkgList().

Examples

biocMaintained()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> # A tibble: 55 × 46
#>    Package        Version Depends Suggests License MD5sum NeedsCompilation Title
#>    <chr>          <chr>   <list>  <list>   <chr>   <chr>  <chr>            <chr>
#>  1 annotate       1.72.0  <chr>   <chr>    Artist… 6ebfd… no               "Ann…
#>  2 AnnotationDbi  1.56.2  <chr>   <chr>    Artist… 9cda1… no               "Man…
#>  3 AnnotationFil… 1.18.0  <chr>   <chr>    Artist… 3bd3d… no               "Fac…
#>  4 AnnotationFor… 1.36.0  <chr>   <chr>    Artist… b888e… no               "Too…
#>  5 AnnotationHub  3.2.2   <chr>   <chr>    Artist… 15045… yes              "Cli…
#>  6 AnnotationHub… 1.24.2  <chr>   <chr>    Artist… e8c3f… no               "Tra…
#>  7 Biobase        2.54.0  <chr>   <chr>    Artist… bd29c… yes              "Bio…
#>  8 BiocCheck      1.30.0  <chr>   <chr>    Artist… 796ea… no               "Bio…
#>  9 BiocDockerMan… 1.6.0   <chr>   <chr>    Artist… ca01f… no               "Acc…
#> 10 BiocGenerics   0.40.0  <chr>   <chr>    Artist… cf429… no               "S4 …
#> # … with 45 more rows, and 38 more variables: Description <chr>,
#> #   biocViews <list>, Author <list>, Maintainer <list>, git_url <chr>,
#> #   git_branch <chr>, git_last_commit <chr>, git_last_commit_date <chr>,
#> #   `Date/Publication` <chr>, source.ver <chr>, win.binary.ver <chr>,
#> #   mac.binary.ver <chr>, vignettes <list>, vignetteTitles <list>,
#> #   hasREADME <chr>, hasNEWS <chr>, hasINSTALL <chr>, hasLICENSE <chr>,
#> #   Rfiles <list>, dependencyCount <chr>, Imports <list>, Enhances <list>, …