Get full bioc software package listing, with details

The BiocViews-generated VIEWS file is available for bioconductor release and devel repositories. It contains quite a bit more information from the package DESCRIPTION files than the PACKAGES file. In particular, it contains biocViews annotations and URLs for vignettes and developer URLs.

biocPkgList(version = BiocManager::version(), repo = "BioCsoft")

Arguments

version
The requested bioconductor version. Will default to use the BiocManager defaults (ie., version()).

repo
The requested biooconductor repository. The default will be the Bioconductor software repository: BioCsoft. Available repos include: "BioCsoft", "BioCann", "BioCexp", "BioCworkflows", and "CRAN". Note that not all repos are available for all versions, particularly older versions (but who would use those, right?).

Value

an object of class tbl_df.

Examples

bpkgl = biocPkgList("3.7")
bpkgl

#> # A tibble: 1,562 x 43
#>    Package Version Depends Suggests License MD5sum NeedsCompilation Title
#>    <fct>   <fct>   <list>  <list>   <fct>   <fct>  <fct>            <fct>
#>  1 a4      1.28.0  <chr [… <chr [4… GPL-3   44fdf… no               Auto…
#>  2 a4Base  1.28.0  <chr [… <chr [2… GPL-3   20974… no               Auto…
#>  3 a4Clas… 1.28.0  <chr [… <chr [1… GPL-3   61aae… no               Auto…
#>  4 a4Core  1.28.0  <chr [… <chr [1… GPL-3   0b686… no               Auto…
#>  5 a4Prep… 1.28.0  <chr [… <chr [2… GPL-3   192fa… no               Auto…
#>  6 a4Repo… 1.28.0  <chr [… <chr [1… GPL-3   2682b… no               Auto…
#>  7 ABAEnr… 1.10.0  <chr [… <chr [3… GPL (>… c0817… yes              Gene…
#>  8 ABarray 1.48.0  <chr [… <chr [2… GPL     aa30d… no               "Mic…
#>  9 ABSSeq  1.34.1  <chr [… <chr [1… GPL (>… f6367… no               "ABS…
#> 10 acde    1.10.0  <chr [… <chr [2… GPL-3   f6784… no               "Art…
#> # ... with 1,552 more rows, and 35 more variables: Description <fct>,
#> #   biocViews <fct>, Author <list>, Maintainer <list>, source.ver <fct>,
#> #   win.binary.ver <fct>, `mac.binary.el-capitan.ver` <fct>, vignettes <list>,
#> #   vignetteTitles <list>, hasREADME <fct>, hasNEWS <fct>, hasINSTALL <fct>,
#> #   hasLICENSE <fct>, Rfiles <fct>, Enhances <list>, dependsOnMe <list>,
#> #   Imports <list>, importsMe <list>, suggestsMe <list>, LinkingTo <fct>,
#> #   Archs <fct>, VignetteBuilder <fct>, git_url <fct>, git_branch <fct>,
#> #   git_last_commit <fct>, git_last_commit_date <fct>,
#> #   `Date/Publication` <fct>, URL <fct>, BugReports <fct>,
#> #   SystemRequirements <fct>, Video <fct>, OS_type <fct>,
#> #   License_restricts_use <fct>, PackageStatus <fct>, License_is_FOSS <fct>

unlist(bpkgl[1,'Depends'])

#>      Depends1      Depends2      Depends3      Depends4      Depends5 
#>      "a4Base"   "a4Preproc"   "a4Classif"      "a4Core" "a4Reporting" 

# Get a list of all packages that
# import "GEOquery"
library(dplyr)

#> Warning: package ‘dplyr’ was built under R version 3.5.1

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

bpkgl %>%
  filter(Package=='GEOquery') %>%
  pull(c('importsMe'))

#> [[1]]
#>  [1] "AnnotationHubData" "bigmelon"          "ChIPXpress"       
#>  [4] "coexnet"           "crossmeta"         "EGAD"             
#>  [7] "GSEABenchmarkeR"   "MACPET"            "minfi"            
#> [10] "MoonlightR"        "phantasus"         "recount"          
#> [13] "SRAdb"            
#>