Get full Bioconductor software package listing, with details

The BiocViews-generated VIEWS file is available for Bioconductor release and devel repositories. It contains quite a bit more information from the package DESCRIPTION files than the PACKAGES file. In particular, it contains biocViews annotations and URLs for vignettes and developer URLs.

biocPkgList(
  version = BiocManager::version(),
  repo = "BioCsoft",
  addBiocViewParents = TRUE
)

Arguments

version
The requested Bioconductor version. Will default to use the BiocManager defaults (ie., version()).

repo
The requested biooconductor repository. The default will be the Bioconductor software repository: BioCsoft. Available repos include: "BioCsoft", "BioCann", "BioCexp", "BioCworkflows", and "CRAN". Note that not all repos are available for all versions, particularly older versions (but who would use those, right?).

addBiocViewParents
logical(), whether to add all biocViews parents to biocViews annotations.

Value

An object of class tbl_df.

Details

Since packages are annotated with the most specific views, the default functionality here is to add parent terms for all views for each package. For example, in the bioCsoft repository, all packages will have at least "Software" added to their biocViews. If one wants to stick to only the most specific terms, set addBiocViewParents to FALSE.

Examples

bpkgl = biocPkgList()
bpkgl

#> # A tibble: 1,838 x 46
#>    Package Version Depends Suggests License MD5sum NeedsCompilation Title
#>    <chr>   <chr>   <list>  <list>   <chr>   <chr>  <chr>            <chr>
#>  1 a4      1.35.0  <chr [… <chr [4… GPL-3   5d200… no               "Aut…
#>  2 a4Base  1.35.0  <chr [… <chr [2… GPL-3   b8ce9… no               "Aut…
#>  3 a4Clas… 1.35.0  <chr [… <chr [1… GPL-3   652e9… no               "Aut…
#>  4 a4Core  1.35.0  <chr [… <chr [1… GPL-3   9543f… no               "Aut…
#>  5 a4Prep… 1.35.0  <chr [… <chr [2… GPL-3   412f4… no               "Aut…
#>  6 a4Repo… 1.35.0  <chr [… <chr [1… GPL-3   6382c… no               "Aut…
#>  7 ABAEnr… 1.17.0  <chr [… <chr [3… GPL (>… 1f587… yes              "Gen…
#>  8 ABarray 1.55.0  <chr [… <chr [2… GPL     d23ec… no               "Mic…
#>  9 abseqR  1.5.0   <chr [… <chr [1… GPL-3 … 17238… no               "Rep…
#> 10 ABSSeq  1.41.0  <chr [… <chr [1… GPL (>… 2f772… no               "ABS…
#> # … with 1,828 more rows, and 38 more variables: Description <chr>,
#> #   biocViews <list>, Author <list>, Maintainer <list>, git_url <chr>,
#> #   git_branch <chr>, git_last_commit <chr>, git_last_commit_date <chr>,
#> #   `Date/Publication` <chr>, source.ver <chr>, win.binary.ver <chr>,
#> #   `mac.binary.el-capitan.ver` <chr>, vignettes <list>, vignetteTitles <list>,
#> #   hasREADME <chr>, hasNEWS <chr>, hasINSTALL <chr>, hasLICENSE <chr>,
#> #   Rfiles <list>, dependencyCount <chr>, Enhances <list>, dependsOnMe <list>,
#> #   Imports <list>, importsMe <list>, suggestsMe <list>, LinkingTo <list>,
#> #   Archs <list>, VignetteBuilder <chr>, URL <chr>, SystemRequirements <chr>,
#> #   BugReports <chr>, Video <chr>, linksToMe <list>, OS_type <chr>,
#> #   PackageStatus <chr>, License_restricts_use <chr>, License_is_FOSS <chr>,
#> #   organism <chr>

unlist(bpkgl[1,'Depends'])

#>      Depends1      Depends2      Depends3      Depends4      Depends5 
#>      "a4Base"   "a4Preproc"   "a4Classif"      "a4Core" "a4Reporting" 

# Get a list of all packages that
# import "GEOquery"
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

bpkgl %>%
  filter(Package=='GEOquery') %>%
  pull(c('importsMe'))

#> [[1]]
#>  [1] "bigmelon"   "ChIPXpress" "coexnet"    "crossmeta"  "EGAD"      
#>  [6] "GAPGOM"     "MACPET"     "minfi"      "MoonlightR" "phantasus" 
#> [11] "recount"    "SRAdb"     
#>