Notify downstream maintainers of changes in upstream package

The `biocRevDepEmail` function collects all the emails of the reverse dependencies and sends a notification that an upstream package has been deprecated or removed. It uses a template found in `inst/resources` with the `templatePath()` function.

biocRevDepEmail(
  pkg,
  which = c("strong", "most", "all"),
  PS = character(1L),
  version = BiocManager::version(),
  dry.run = TRUE,
  cc = NULL,
  emailTemplate = templatePath("revdepnote"),
  core.name = NULL,
  core.email = NULL,
  core.id = NULL,
  textOnly = FALSE,
  verbose = FALSE,
  credFile = "~/.blastula_creds"
)

Arguments

pkg

character(1) The name of the package for whose reverse dependencies are to be checked and notified.

which

a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

PS

character(1) Postscript, an additional note to the recipient of the email (i.e., the package maintainer)

version

character() A vector indicating which version of Bioconductor the package is failing in (either 'release' or 'devel'; defaults to both)

dry.run

logical(1) Display the email without sending to the recipient. It only works for HTML email reports and ignored when `textOnly=TRUE`

cc

character() Email addresses for sending the message as a carbon copy

emailTemplate

character(1) The path to the email template Rmd file as obtained by `templatePath()`. A custom template can be provided as file path.

core.name

character(1) The full name of the core team member

core.email

character(1) The Roswell Park email of the core team member

core.id

character(1) The internal identifier for the Roswell employee. This ID usually matches `^[A-Z]2[0-9]5` for more recent identifiers.

textOnly

logical(1) Whether to return the text of the email only. This avoids the use of the 'blastula' package and adds the text to the system clipboard if the `clipr` package is installed (default: FALSE)

verbose

logical(1) Whether to output full email information from 'smtp_send' (when `dry.run` is `FALSE` and 'blastula' is installed)

credFile

character(1) An optional file generated by the `blastula::create_smtp_creds_file` function containing email authentication information (default: "~/.blastula_creds"). See `?biocBuildEmail` details.

Examples

biocRevDepEmail(
    "FindMyFriends", version = "3.13", dry.run = TRUE, textOnly = TRUE
)
#> Provide your full name: 
#> What is your core-team email? 
#> What is your employee ID (e.g., AB12345, leave blank if N/A)? 
#> Saved data to: /home/runner/.cache/R/BiocPkgTools
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org
#> Message not sent: Set 'dry.run=FALSE'
#> [1] "thomasp85@gmail.com\nBioconductor Package FindMyFriends Deprecation Notification\n\n\n\nHi Maintainers,\n\nUnfortunately, due to the deprecation of the `FindMyFriends` package in Bioconductor,\nyour package has been affected. Please remove the facilities of the deprecated\npackage from your package. This includes the removal of imported functions and\nthe package from your `DESCRIPTION` file.\n\nHere is the full list of affected packages:\n\n[PanVizGenerator](https://bioconductor.org/checkResults/3.13/bioc-LATEST/PanVizGenerator)\n\nWe apologize for this inconvenience and if you have any questions please\ndirect them to the mailing list at bioc-devel@bioconductor.org.\n\nhttps://bioconductor.org/checkResults/3.13/bioc-LATEST/FindMyFriends\n\nBest regards,\n\nNA\n\n\n\n\n"