Work with package dependencies as a graph

Package dependencies represent a directed graph (though Bioconductor dependencies are not an acyclic graph). This function simply returns an igraph graph from the package dependency data frame from a call to buildPkgDependencyDataFrame or any tidy data frame with rows of (Package, dependency) pairs. Additional columns are added as igraph edge attributes (see graph_from_data_frame).

buildPkgDependencyIgraph(pkgDepDF)

Arguments

pkgDepDF
a tidy data frame. See description for details.

Value

an igraph directed graph. See the igraph package for details of what can be done.

See also

Examples

library(igraph)

#> 
#> Attaching package: ‘igraph’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     as_data_frame, groups, union

#> The following object is masked from ‘package:testthat’:
#> 
#>     compare

#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum

#> The following object is masked from ‘package:base’:
#> 
#>     union

pkg_dep_df = buildPkgDependencyDataFrame()

#> Warning: data set ‘biocViewsVocab’ not found

#> Error in nodes(g): object 'biocViewsVocab' not found

# at this point, filter or join to manipulate
# dependency data frame as you see fit.

g = buildPkgDependencyIgraph(pkg_dep_df)

#> Error in as.data.frame(d): object 'pkg_dep_df' not found

g

#> Error in eval(expr, envir, enclos): object 'g' not found

# Look at nodes and edges
head(V(g)) # vertices

#> Error in "igraph" %in% class(graph): object 'g' not found

head(E(g)) # edges

#> Error in "igraph" %in% class(graph): object 'g' not found

# subset graph by attributes

head(sort(degree(g, mode='in'), decreasing=TRUE))

#> Error in "igraph" %in% class(graph): object 'g' not found

head(sort(degree(g, mode='out'), decreasing=TRUE))

#> Error in "igraph" %in% class(graph): object 'g' not found