Work with package dependencies as a graph

Package dependencies represent a directed graph (though Bioconductor dependencies are not an acyclic graph). This function simply returns an igraph graph from the package dependency data frame from a call to buildPkgDependencyDataFrame or any tidy data frame with rows of (Package, dependency) pairs. Additional columns are added as igraph edge attributes (see graph_from_data_frame).

buildPkgDependencyIgraph(pkgDepDF)

Arguments

pkgDepDF

a tidy data frame. See description for details.

Value

An igraph directed graph. See the igraph package for details of what can be done.

See also

Examples

library(igraph)
#> 
#> Attaching package: ‘igraph’
#> The following objects are masked from ‘package:dplyr’:
#> 
#>     as_data_frame, groups, union
#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum
#> The following object is masked from ‘package:base’:
#> 
#>     union

pkg_dep_df = buildPkgDependencyDataFrame()
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> '?repositories' for details
#> 
#> replacement repositories:
#>     CRAN: https://cloud.r-project.org

# at this point, filter or join to manipulate
# dependency data frame as you see fit.

g = buildPkgDependencyIgraph(pkg_dep_df)
g
#> IGRAPH 302a0f0 DN-- 3752 36433 -- 
#> + attr: name (v/c), edgetype (e/c)
#> + edges from 302a0f0 (vertex names):
#>  [1] a4           ->a4Base      a4           ->a4Preproc  
#>  [3] a4           ->a4Classif   a4           ->a4Core     
#>  [5] a4           ->a4Reporting a4Base       ->a4Preproc  
#>  [7] a4Base       ->a4Core      a4Classif    ->a4Core     
#>  [9] a4Classif    ->a4Preproc   ABAEnrichment->R          
#> [11] abseqR       ->R           ABSSeq       ->R          
#> [13] ABSSeq       ->methods     acde         ->R          
#> [15] acde         ->boot        ACE          ->R          
#> + ... omitted several edges

# Look at nodes and edges
head(V(g)) # vertices
#> + 6/3752 vertices, named, from 302a0f0:
#> [1] a4            a4Base        a4Classif     ABAEnrichment abseqR       
#> [6] ABSSeq       
head(E(g)) # edges
#> + 6/36433 edges from 302a0f0 (vertex names):
#> [1] a4    ->a4Base      a4    ->a4Preproc   a4    ->a4Classif  
#> [4] a4    ->a4Core      a4    ->a4Reporting a4Base->a4Preproc  

# subset graph by attributes

head(sort(degree(g, mode='in'), decreasing=TRUE))
#>         R     knitr   methods rmarkdown     stats     utils 
#>      1653      1419      1408      1227      1185       982 
head(sort(degree(g, mode='out'), decreasing=TRUE))
#> singleCellTK     musicatk TCGAbiolinks   ChromSCape ChIPpeakAnno    metaseqR2 
#>           94           64           64           62           61           61