Work with package dependencies as a graph

Package dependencies represent a directed graph (though Bioconductor dependencies are not an acyclic graph). This function simply returns an igraph graph from the package dependency data frame from a call to buildPkgDependencyDataFrame or any tidy data frame with rows of (Package, dependency) pairs. Additional columns are added as igraph edge attributes (see graph_from_data_frame).

buildPkgDependencyIgraph(pkgDepDF)

Arguments

pkgDepDF
a tidy data frame. See description for details.

Value

an igraph directed graph. See the igraph package for details of what can be done.

See also

Examples

library(igraph)

#> 
#> Attaching package: ‘igraph’

#> The following objects are masked from ‘package:dplyr’:
#> 
#>     as_data_frame, groups, union

#> The following objects are masked from ‘package:stats’:
#> 
#>     decompose, spectrum

#> The following object is masked from ‘package:base’:
#> 
#>     union

pkg_dep_df = buildPkgDependencyDataFrame()

# at this point, filter or join to manipulate
# dependency data frame as you see fit.

g = buildPkgDependencyIgraph(pkg_dep_df)
g

#> IGRAPH cce2816 DN-- 2796 21840 -- 
#> + attr: name (v/c), edgetype (e/c)
#> + edges from cce2816 (vertex names):
#>  [1] a4       ->a4Base        a4       ->a4Preproc     a4       ->a4Classif    
#>  [4] a4       ->a4Core        a4       ->a4Reporting   a4Base   ->methods      
#>  [7] a4Base   ->graphics      a4Base   ->grid          a4Base   ->Biobase      
#> [10] a4Base   ->AnnotationDbi a4Base   ->annaffy       a4Base   ->mpm          
#> [13] a4Base   ->genefilter    a4Base   ->limma         a4Base   ->multtest     
#> [16] a4Base   ->glmnet        a4Base   ->a4Preproc     a4Base   ->a4Core       
#> [19] a4Base   ->gplots        a4Classif->methods       a4Classif->a4Core       
#> [22] a4Classif->a4Preproc     a4Classif->MLInterfaces  a4Classif->ROCR         
#> + ... omitted several edges

# Look at nodes and edges
head(V(g)) # vertices

#> + 6/2796 vertices, named, from cce2816:
#> [1] a4          a4Base      a4Classif   a4Core      a4Preproc   a4Reporting

head(E(g)) # edges

#> + 6/21840 edges from cce2816 (vertex names):
#> [1] a4    ->a4Base      a4    ->a4Preproc   a4    ->a4Classif  
#> [4] a4    ->a4Core      a4    ->a4Reporting a4Base->methods    

# subset graph by attributes

head(sort(degree(g, mode='in'), decreasing=TRUE))

#>       R methods   knitr   stats   utils Biobase 
#>    1206    1000     717     675     570     546 

head(sort(degree(g, mode='out'), decreasing=TRUE))

#>    TCGAbiolinks         RnBeads BiocCaseStudies    ChIPpeakAnno         MSnbase 
#>              65              57              52              52              49 
#>       debrowser 
#>              48