Generate a DOI for a bioconductor package

This function makes calls out to the EZID API (v2) described here: https://ezid.lib.purdue.edu/doc/apidoc.2.html. The function creates a new DOI for a bioc package (cannot already exist). The target URL for the DOI is the short Bioconductor package URL.

generateBiocPkgDOI(pkg, authors, pubyear, testing = TRUE)

Arguments

pkg
character(1) package name

authors
character vector of authors (will be "pasted" together)

pubyear
integer(1) publication year

testing
logical(1) If true, will use the apitest user with the password apitest. These DOIs will expire. The same apitest:apitest combination can be used to login to the EZID website for doing things using the web interface. If false, the Bioconductor-specific user credentials should be in the correct environment variables

Value

The DOI as a character(1) vector.

Details

The login information for the "real" Bioconductor account should be stored in the environment variables "EZID_USERNAME" and "EZID_PASSWORD".

The GUI is available here: https://ezid.lib.purdue.edu.

Examples

# NOT RUN {
  x = generateBiocPkgDOI('RANDOM_TEST_PACKAGE','Sean Davis',1972)
# }