Get package details from GitHub

For packages that live on github, we can mine further details. This function returns the github details for the listed packages.

githubDetails(pkgs, sleep = 0)

Arguments

pkgs
a character() vector of username/repo for one or more github repos, such as `seandavi/GEOquery`.

sleep
numeric() denoting the number of seconds to sleep between GitHub API calls. Since GitHub rate limits its APIs, it might be necessary to either use small chunks of packages iteratively or to supply a non-zero argument here. See the `details` section for a better solution using GitHub tokens.

Details

The gh function is used to do the fetching. If the number of packages supplied to this function is large (>40 or so), it is possible to run into problems with API rate limits. The gh package uses the environment variable "GITHUB_PAT" (for personal access token) to authenticate and then provide higher rate limits. If you run into problems with rate limits, set sleep to some small positive number to slow queries. Alternatively, create a Personal Access Token on GitHub and register it. See the gh package for details.

Examples

pkglist = biocPkgList()

#> Warning: data set ‘biocViewsVocab’ not found

#> Error in nodes(g): object 'biocViewsVocab' not found

# example of "pkgs" format.
head(pkglist$URL)

#> Error in head(pkglist$URL): object 'pkglist' not found

gh_list = githubURLParts(pkglist$URL)

#> Error in stri_match_first_regex(string, pattern, opts_regex = opts(pattern)): object 'pkglist' not found

gh_list = gh_list[!is.null(gh_list$user_repo),]

#> Error in eval(expr, envir, enclos): object 'gh_list' not found

head(gh_list$user_repo)

#> Error in head(gh_list$user_repo): object 'gh_list' not found

ghd = githubDetails(gh_list$user_repo[1:5])

#> Error in lapply(pkgs, function(pkg) {    Sys.sleep(sleep)    .gh_pkg_info(pkg)}): object 'gh_list' not found

lapply(ghd, '[[', "stargazers")

#> Error in lapply(ghd, "[[", "stargazers"): object 'ghd' not found