Look up a package's Bioconductor dependencies

The function uses the pkgType argument to restrict the look up to only the relevant Bioconductor repository. It works for multiple packages of the same type.

pkgBiocDeps(
  pkg,
  pkgType = c("software", "data-experiment", "workflows", "data-annotation"),
  which = "strong",
  only.bioc = TRUE,
  recursive = FALSE,
  version = BiocManager::version()
)

Arguments

pkg

character(1) The package for which to look up dependencies.

pkgType

character() Any of 'software', 'data-experiment', 'workflows', and / or 'data-annotation' (defaults to all)

which

a character vector listing the types of dependencies, a subset of c("Depends", "Imports", "LinkingTo", "Suggests", "Enhances"). Character string "all" is shorthand for that vector, character string "most" for the same vector without "Enhances", character string "strong" (default) for the first three elements of that vector.

only.bioc

logical(1) Whether to only return Bioconductor dependencies in the list (default TRUE)

recursive

a logical indicating whether (reverse) dependencies of (reverse) dependencies (and so on) should be included, or a character vector like which indicating the type of (reverse) dependencies to be added recursively.

version

(Optional) character(1) or package_version indicating the Bioconductor version (e.g., "3.8") for which repositories are required.

Examples

pkgBiocDeps("MultiAssayExperiment", only.bioc = TRUE)
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
#> $MultiAssayExperiment
#> [1] "SummarizedExperiment" "Biobase"              "BiocBaseUtils"       
#> [4] "BiocGenerics"         "DelayedArray"         "GenomicRanges"       
#> [7] "IRanges"              "S4Vectors"           
#> 

pkgBiocDeps("MultiAssayExperiment", only.bioc = FALSE)
#> 'getOption("repos")' replaces Bioconductor standard repositories, see
#> 'help("repositories", package = "BiocManager")' for details.
#> Replacement repositories:
#>     CRAN: https://cran.rstudio.com
#> $MultiAssayExperiment
#>  [1] "SummarizedExperiment" "Biobase"              "BiocBaseUtils"       
#>  [4] "BiocGenerics"         "DelayedArray"         "GenomicRanges"       
#>  [7] "IRanges"              "methods"              "S4Vectors"           
#> [10] "tidyr"                "utils"               
#>