Calculate dependency gain achieved by excluding combinations of packages

Calculate dependency gain achieved by excluding combinations of packages

pkgCombDependencyGain(pkg, depdf, maxNbr = 3L)

Arguments

pkg
character, the name of the package for which we want to estimate the dependency gain

depdf
a tidy data frame with package dependency information obtained through the function buildPkgDependencyDataFrame

maxNbr
numeric, the maximal number of direct dependencies to leave out simultaneously

Value

A data frame with three columns: ExclPackages (the excluded direct dependencies), NbrExcl (the number of excluded direct dependencies), DepGain (the dependency gain from excluding these direct dependencies)

Examples

depdf <- buildPkgDependencyDataFrame(
  dependencies=c("Depends", "Imports"),
  repo=c("BioCsoft", "CRAN")
)
pcd <- pkgCombDependencyGain('GEOquery', depdf, maxNbr = 3L)
head(pcd[order(pcd$DepGain, decreasing = TRUE), ])

#>                Packages NbrExcl DepGain
#> 77  dplyr, readr, tidyr       3      22
#> 48 Biobase, httr, readr       3      13
#> 49 Biobase, httr, tidyr       3      13
#> 74   dplyr, httr, tidyr       3      13
#> 83   httr, readr, tidyr       3      13
#> 38 Biobase, limma, httr       3      12